backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Hebo backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-allauth>=65.3.0",
    "django>=5.1.4",
    "django-organizations>=2.5.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "pyjwt>=2.10.1",
    "cryptography>=44.0.0",
    "markdown>=3.7",
    "djangorestframework>=3.15.2",
    "pgvector>=0.3.6",
    "django-extensions>=3.2.3",
    "django-schema-graph>=3.1.0",
    "dj-stripe>=2.8.4",
    "django-stubs>=5.1.1",
]
